Update terms-and-conditions.md
In '2. Acceptance' the following was added:
"When you make an order for the charging of an electric vehicle through the Services, Monta acts as an agent on behalf of the Operator to conclude your order and to manage your experience throughout the order process. Monta is not liable and does not take any responsibility for the provision of the charging of electric vehicles by users."